modify service-workflow demo to follow go 101 example
