+++ demos/service-workflow/README.md
@@ -18,27 +18,20 @@ Since this Workflow depends on the translation microservice, start that
 first by running this command in a terminal window:
 
 ```command
-npm run service.watch
+npm run service
 ```
 
 Next, use a separate terminal window to start the Worker:
 
 ```command
-npm run worker.watch
+npm run worker
 ```
 
-Finally, use another terminal window to start the Workflows. You can change the name passed to the greeting Workflow in the `src/clients/greeting.ts` file and to the farewell Workflow in the `src/clients/farewell.ts` file.
-
-To start the `greeting` Workflow, run the command:
-
-```command
-npm run greeting
-```
-
-To start the `farewell` Workflow, run the command:
+inally, use another terminal window to start the Workflow,
+using the supplied code, specifying your name as input.
 
 ```command
-npm run farewell
+npm run workflow Donna
 ```
 
 #### This is the end of the demo
