+++ demos/service-workflow/src/activities.ts
@@ -7,9 +7,3 @@ export async function getSpanishGreeting(name: string): Promise<string> {
 
   return response.data;
 }
-
-export async function getSpanishFarewell(name: string): Promise<string> {
-  const response = await axios.get(`${url}/get-spanish-farwell?name=${name}`);
-
-  return response.data;
-}
